refactor(stage-ui): organize Live2D motion devtools
