+++ apps/stage-tamagotchi/package.json
@@ -65,7 +65,6 @@
     "@proj-airi/i18n": "workspace:^",
     "@proj-airi/input-gamepad": "workspace:^",
     "@proj-airi/model-driver-lipsync": "workspace:^",
-    "@proj-airi/model-live2d-motion": "workspace:^",
     "@proj-airi/pipelines-audio": "workspace:^",
     "@proj-airi/plugin-sdk-tamagotchi": "workspace:^",
     "@proj-airi/server-sdk": "workspace:*",
@@ -105,11 +104,9 @@
     "chess.js": "catalog:",
     "colorjs.io": "catalog:",
     "culori": "catalog:",
-    "d3": "catalog:",
     "date-fns": "catalog:",
     "defu": "catalog:",
     "destr": "catalog:",
-    "dockview-vue": "catalog:",
     "dompurify": "catalog:",
     "electron-click-drag-plugin": "catalog:",
     "embla-carousel-vue": "catalog:",
@@ -185,7 +182,6 @@
     "@proj-airi/unplugin-live2d-sdk": "catalog:",
     "@types/audioworklet": "catalog:",
     "@types/culori": "catalog:",
-    "@types/d3": "catalog:",
     "@types/nprogress": "catalog:",
     "@types/semver": "catalog:",
     "@types/splitpanes": "catalog:",
