+++ packages/stage-ui/package.json
@@ -27,6 +27,7 @@
     "./constants/*": "./src/constants/*.ts",
     "./constants": "./src/constants/index.ts",
     "./directives/*": "./src/directives/*.ts",
+    "./features/devtools/motion/live2d": "./src/features/devtools/motion/live2d/index.ts",
     "./libs/inference/adapters/*": "./src/libs/inference/adapters/*.ts",
     "./libs/inference": "./src/libs/inference/index.ts",
     "./libs/analytics": "./src/libs/analytics/index.ts",
@@ -88,7 +89,9 @@
     "@proj-airi/font-cjkfonts-allseto": "workspace:^",
     "@proj-airi/font-xiaolai": "workspace:^",
     "@proj-airi/i18n": "workspace:^",
+    "@proj-airi/input-gamepad": "workspace:^",
     "@proj-airi/model-driver-lipsync": "workspace:^",
+    "@proj-airi/model-live2d-motion": "workspace:^",
     "@proj-airi/pipelines-audio": "workspace:^",
     "@proj-airi/server-sdk": "workspace:^",
     "@proj-airi/server-sdk-shared": "workspace:^",
@@ -127,6 +130,7 @@
     "culori": "catalog:",
     "d3": "catalog:",
     "date-fns": "catalog:",
+    "dockview-vue": "catalog:",
     "dompurify": "catalog:",
     "embla-carousel-autoplay": "catalog:",
     "embla-carousel-vue": "catalog:",
