+++ packages/stage-ui/src/features/devtools/motion/live2d/assets/idle-excited.json
