+++ packages/stage-ui/src/features/devtools/motion/live2d/components/preview.vue
