+++ packages/stage-ui/src/features/devtools/motion/live2d/components/timeline-ruler.vue
