+++ packages/stage-ui/src/features/devtools/motion/live2d/composables/default-recording.ts
@@ -1,6 +1,6 @@
-import idleExcitedProjectJson from '../assets/devtools/live2d-motion/idle-excited.json?raw'
+import idleExcitedProjectJson from '../assets/idle-excited.json?raw'
 
-import { parseLive2DMotionProject } from './live2d-motion-keyframes'
+import { parseLive2DMotionProject } from './keyframes'
 
 /** The validated trajectory that initializes the Live2D motion devtool. */
 export const defaultLive2DMotionRecording = parseLive2DMotionProject(idleExcitedProjectJson).source
