+++ packages/stage-ui/src/features/devtools/motion/live2d/composables/recording.ts
