+++ packages/stage-ui/src/features/devtools/motion/live2d/index.ts
@@ -1,0 +1,2 @@
+/** Interactive Live2D motion workbench for shared stage devtools. */
+export { default as Live2DMotionDevtools } from './devtools.vue'
