+++ packages/stage-ui/vitest.config.ts
@@ -1,4 +1,3 @@
-import { join } from 'node:path'
 import { cwd } from 'node:process'
 
 import Vue from '@vitejs/plugin-vue'
@@ -8,47 +7,43 @@ import { playwright } from '@vitest/browser-playwright'
 import { loadEnv } from 'vite'
 import { defineConfig } from 'vitest/config'
 
-export default defineConfig(({ mode }) => {
-  return {
-    root: import.meta.dirname,
-    plugins: [
-      Info(),
+export default defineConfig({
+  root: import.meta.dirname,
+  plugins: [
+    Info(),
+    Vue(),
+  ],
+  test: {
+    env: loadEnv('test', cwd(), ''),
+    projects: [
+      {
+        extends: true,
+        test: {
+          name: 'node',
+          include: ['src/**/*.test.ts'],
+          exclude: ['src/**/*.browser.test.ts'],
+          fileParallelism: false,
+          hookTimeout: 20_000,
+          maxWorkers: 1,
+          testTimeout: 20_000,
+        },
+      },
+      {
+        extends: true,
+        test: {
+          name: 'browser',
+          include: ['src/**/*.browser.test.ts'],
+          exclude: ['**/node_modules/**'],
+          browser: {
+            enabled: true,
+            headless: true,
+            provider: playwright(),
+            instances: [
+              { browser: 'chromium' },
+            ],
+          },
+        },
+      },
     ],
-    test: {
-      projects: [
-        {
-          extends: true,
-          test: {
-            name: 'node',
-            include: ['src/**/*.test.ts'],
-            exclude: ['src/**/*.browser.test.ts'],
-            env: loadEnv(mode, join(cwd(), 'packages', 'stage-ui'), ''),
-            fileParallelism: false,
-            hookTimeout: 20_000,
-            maxWorkers: 1,
-            testTimeout: 20_000,
-          },
-        },
-        {
-          extends: true,
-          plugins: [
-            Vue(),
-          ],
-          test: {
-            name: 'browser',
-            include: ['**/*.browser.{spec,test}.ts'],
-            exclude: ['**/node_modules/**'],
-            browser: {
-              enabled: true,
-              headless: true,
-              provider: playwright(),
-              instances: [
-                { browser: 'chromium' },
-              ],
-            },
-          },
-        },
-      ],
-    },
-  }
+  },
 })
